[UPD] run.py --> fix bug #12 [UPD] config.ini.tutorial --> fix bug #12
The user can store a fixed time in the config.ini file. If this time is
expired while the television was in 'On' state, a reboot of the television
is performed. If the television is set to 'Off' and later on back to 'On'
state, the counter start again at the time of the last 'On' state order.